+++ Scheduling Algorithms/SchedStats.hpp
@@ -7,7 +7,7 @@ struct SchedStats {
 	std::string AlgoUsed;
 	double avgWaitTime;
 	double avgTurnAroundTime;
-	size_t maxWaitTime;
-	size_t maxTurnAroundTime;
+	long maxWaitTime;
+	long maxTurnAroundTime;
 	std::vector<GanttNode> ganttChart;
 };
